requirements.txt
Django==3.0.6
django-model-utils==4.0.0
